pyproject.toml
[project]
name = "gemini_examples"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.5",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "langchainhub>=0.1.14",
    "langgraph>=0.2.45",
    "langgraph-swarm>=0.0.11",
    "cryptography>=41.0.0",
]
